CreateView Controller modified to include character count and views are correct auto layout using iPhone 4s onwards

+++ EditCellCollectionViewCell.m
@@ -86,14 +86,13 @@
 }
 
 - (void)textViewDidChange:(UITextView *)textView    {
-    _charCount.text =  [NSString stringWithFormat:@"%d", textView.text.length];
+    _charCount.text =  [NSString stringWithFormat:@"%lu", (unsigned long)textView.text.length];
 
     if (textView.text.length != 140) {
         _charCount.textColor = [UIColor blueColor];
     } else {
         _charCount.textColor = [UIColor redColor];
     }
-
 }
 
 #pragma mark - Emoji Methods
